WORKING CODE.
Changes:
- Added ERROR in enum MessageType to handle errors in the future (ex. Database is down).
- Created new class NetworkHandler and now using that instead of Network.sendRequest.
Reason is cause I wanted to separate the reading and writing.
Also suspecting another problem with Network.sendRequest: When you're trying to ACK
you don't want to read back, so you can't use it. And in cases where you already have
a Socket (ex. serverSocket.accept()) you have to take it as it is.
- Made Checkin implement Serializable for Network transmission.
- Changed from Collections.toList() to Collections.toCollection(ArrayList::new)
to have more control over what is returned, make sure the types after being
serialized are the same.

Info:
At first I thought I had a bug cause all the the data were going to the first
mapper. Then after looking the data, I realised the area I was giving was too big.

Latitude Range: [40.55, 40.98]
Longitude Range: [-74.27, -73.68]

So a good suggested big set of data I tried is:
LowerLeftPoint = 40.55,-74.27
TopRightPoint = 40.98,-73.68
DateTimeStart = 0000-01-01 00:00:00  (birth of the universe)
DateTimeEnd = 2022-01-01 00:00:00    (when we'll get a degree)